Initial 7.3 Changes (BROKENish): [PUSH]
Things are still broken on the Dalamud API side of things, do not expect full functionality for a while. Currently cooldown tracking and gauge-related functionality are broken. But hopefully once they do get fixed, this should work.

Also, made Flare Star not get treated as a single-target spell for the purpose of AoE switch stuff.